moved SimpleTcpSvc implementation code to the main reactor directory
